Restructured project to live under ~/.vim
Also updated the installation script a whole lot

+++ install.sh
@@ -8,7 +8,6 @@ set -eof pipefail
 
 # Set global variables
 backup_dir="${HOME}/.vinter-backup.$(date +%Y%m%dT%H%M%S)"
-force_install=0
 source_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 target_dir=${HOME}
 
@@ -23,7 +22,6 @@ print_usage()
 {
   echo -e "Usage: install.sh [OPTIONS]\n"
   echo "OPTIONS:"
-  echo -e "--force\t\tForce install. Will remove your ${target_dir}/.vimrc.local file (will be backed up)"
   echo -e "--help\t\tShow this help"
 }
 
@@ -37,7 +35,7 @@ backup()
     mkdir -p ${backup_dir}
   fi
 
-  if [ -e ${backup_dir}/${1} ] ; then
+  if [ -e ${target_dir}/${1} ] ; then
     echo "Backing up ${target_dir}/${1} to ${backup_dir}/${1}"
     mv ${target_dir}/${1} ${backup_dir}/${1}
   fi
@@ -49,8 +47,26 @@ backup()
 # - File to link to target
 link()
 {
-  echo "Linking ${target_dir}/${1} to ${source_dir}/.${1}"
-  ln -s ${target_dir}/${1} ${source_dir}/.${1}
+  echo "Linking ${source_dir}/${1} to ${target_dir}/.${1}"
+  ln -s ${source_dir}/${1} ${target_dir}/.${1}
+}
+
+# Remove files, dirs or links
+#
+# Args:
+# - Target
+remove()
+{
+  if [ -L ${target_dir}/${1} ] ; then
+    echo "Unlinking ${target_dir}/${1}"
+    unlink ${target_dir}/${1}
+  elif [ -d ${target_dir}/${1} ] ; then
+    echo "Removing ${target_dir}/${1}"
+    rm -rf ${target_dir}/${1}
+  elif [ -f ${target_dir}/${1} ] ; then
+    echo "Removing ${target_dir}/${1}"
+    rm -rf ${target_dir}/${1}
+  fi
 }
 
 ###########################
@@ -62,8 +78,6 @@ link()
 if [ "${1}" == "--help" ] ; then
   print_usage
   exit 0
-elif [ "${1}" == "--force" ] ; then
-  force_install=1
 fi
 
 ###########################
@@ -72,29 +86,35 @@ fi
 #
 ###########################
 
-echo "Installing Vinter"
+echo "> INSTALLING Vinter"
 
 # Backup local vim files
+echo -e "\n>> CREATING BACKUPS"
 echo "Backing up all your old vim files to ${backup_dir}"
 backup ".vim"
 backup ".vimrc"
-backup ".vimrc.plugins"
+backup ".vimrc.local"
+backup ".vimrc.plugins.local"
 
-if [ ${force_install} -eq 1 ] ; then
-  backup ".vimrc.local"
-  backup ".vimrc.plugins.local"
-fi
+echo -e "\n>> CLEAN UP"
+echo "Clean up old vim configuration"
+remove ".vim"
+remove ".vimrc"
 
 # Link vim files
+echo -e "\n>> INSTALL BASE CONFIGURAION"
 echo "Installing Vinter configuration files"
-rm ${target_dir}/.vimrc.local.example > /dev/null 2>&1
 link "vim"
-link "vimrc"
-link "vimrc.plugins"
-link "vimrc.local.example"
-touch ${target_dir}/.vimrc.plugins.local
+
+echo "Creating example .vimrc.local (${target_dir}/.vimrc.local.example)"
+if [ -e ${target_dir}/.vimrc.local.example ] ; then
+  rm ${target_dir}/.vimrc.local.example > /dev/null 2>&1
+fi
+cp ${source_dir}/vimrc.local.example ${target_dir}/.vimrc.local.example
 
 # Install Vundle
+echo -e "\n>> INSTALL DEPENDENCIES AND PLUGINS"
+echo "Installing Vundle for plugin management"
 if [ -d ${target_dir}/.vim/bundle/Vundle.vim ] ; then
     cd ${target_dir}/.vim/bundle/Vundle.vim ; git pull ; cd -
 else
@@ -108,6 +128,7 @@ vim +PluginInstall +qall
 echo "Building the Command T C extension"
 cd ${target_dir}/.vim/bundle/command-t/ruby/command-t && ruby extconf.rb && make
 
+echo -e "\n> DONE"
 echo "Done with the installation. Happy hacking!"
 
 exit 0
